[redirector] Separate counters' functions from records.go
